Quiz exercises: Remove AbstractQuizSubmission and merge into QuizSubmission
AbstractQuizSubmission had only one subclass (QuizSubmission), so the
split added an extra layer without any benefit. Flattens the Java
entity and its mirrored frontend model into QuizSubmission directly,
and updates the handful of call sites (SubmittedAnswer,
SubmittedAnswerRepository, QuizSubmissionService,
AbstractQuizSubmissionService) that referenced the abstract type.
