+++ src/main/java/de/tum/cit/aet/artemis/quiz/domain/SubmittedAnswer.java
@@ -55,7 +55,7 @@ public abstract class SubmittedAnswer extends DomainObject {
 
     @ManyToOne
     @JsonIgnore
-    private AbstractQuizSubmission submission;
+    private QuizSubmission submission;
 
     public Double getScoreInPoints() {
         return scoreInPoints;
@@ -73,11 +73,11 @@ public abstract class SubmittedAnswer extends DomainObject {
         this.quizQuestion = quizQuestion;
     }
 
-    public AbstractQuizSubmission getSubmission() {
+    public QuizSubmission getSubmission() {
         return submission;
     }
 
-    public void setSubmission(AbstractQuizSubmission quizSubmission) {
+    public void setSubmission(QuizSubmission quizSubmission) {
         this.submission = quizSubmission;
     }
