+++ src/main/java/de/tum/cit/aet/artemis/quiz/repository/SubmittedAnswerRepository.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Repository;
 
 import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
 import de.tum.cit.aet.artemis.exercise.domain.participation.StudentParticipation;
-import de.tum.cit.aet.artemis.quiz.domain.AbstractQuizSubmission;
 import de.tum.cit.aet.artemis.quiz.domain.QuizExercise;
 import de.tum.cit.aet.artemis.quiz.domain.QuizSubmission;
 import de.tum.cit.aet.artemis.quiz.domain.SubmittedAnswer;
@@ -24,7 +23,7 @@ import de.tum.cit.aet.artemis.quiz.domain.SubmittedAnswer;
 @Repository
 public interface SubmittedAnswerRepository extends ArtemisJpaRepository<SubmittedAnswer, Long> {
 
-    Set<SubmittedAnswer> findBySubmission(AbstractQuizSubmission submission);
+    Set<SubmittedAnswer> findBySubmission(QuizSubmission submission);
 
     /**
      * Loads submitted answers from the database in case there is a QuizSubmission in one of the passed student participation
