+++ src/main/webapp/app/quiz/shared/entities/quiz-submission.model.ts
@@ -1,7 +1,10 @@
-import { SubmissionExerciseType } from 'app/exercise/shared/entities/submission/submission.model';
-import { AbstractQuizSubmission } from 'app/quiz/shared/entities/abstract-quiz-exam-submission.model';
+import { Submission, SubmissionExerciseType } from 'app/exercise/shared/entities/submission/submission.model';
+import { SubmittedAnswer } from 'app/quiz/shared/entities/submitted-answer.model';
 
-export class QuizSubmission extends AbstractQuizSubmission {
+export class QuizSubmission extends Submission {
+    public scoreInPoints?: number;
+    public submittedAnswers?: SubmittedAnswer[];
+
     constructor() {
         super(SubmissionExerciseType.QUIZ);
     }
